requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
fpdf2>=2.7.0
